Increased doTimeSensitiveProcess parameters' size

+++ src/include/networking.h
@@ -107,7 +107,7 @@ private:
 	unsigned long processStartTime = 0;
 	unsigned short processElapsedTime = 0;
 
-	unsigned short doTimeSensitiveProcess(const unsigned short, const unsigned short, bool (Networking::*)(bool (Networking::*)(Queue<Message>&, QueueNode<Message>*), Queue<Message>&), bool (Networking::*)(Queue<Message>&, QueueNode<Message>*), Queue<Message>&);
+	unsigned short doTimeSensitiveProcess(const unsigned int, const unsigned short, bool (Networking::*)(bool (Networking::*)(Queue<Message>&, QueueNode<Message>*), Queue<Message>&), bool (Networking::*)(Queue<Message>&, QueueNode<Message>*), Queue<Message>&);
 
 	QueueNode<Message>* queueStartNode;
 	QueueNode<Message>* holdingNode;
